memorylayer-sdk-llamaindex-python/pyproject.toml
[project]
name = "memorylayer-llamaindex"
version = "0.0.3"
description = "LlamaIndex integration for MemoryLayer.ai - persistent memory for LlamaIndex applications"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache 2.0" }
authors = [
    { name = "Scitrera", email = "open-source-team@scitrera.com" }
]
keywords = ["ai", "memory", "agents", "llm", "llamaindex", "rag", "vector-search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "memorylayer-client>=0.0.1",
    "llama-index-core>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.20.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://memorylayer.ai"
Documentation = "https://docs.memorylayer.ai"
Repository = "https://github.com/scitrera/memorylayer"
Issues = "https://github.com/scitrera/memorylayer/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
packages = ["src/memorylayer_llamaindex"]

[tool.hatch.build.targets.wheel]
packages = ["src/memorylayer_llamaindex"]

[tool.ruff]
line-length = 140
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
